Move word parser to separate module
