+++ yash-syntax/src/parser/lex.rs
@@ -29,11 +29,7 @@ pub use self::heredoc::PartialHereDoc;
 pub use self::op::is_operator_char;
 
 use self::keyword::Keyword;
-use crate::parser::core::Error;
 use crate::parser::core::Result;
-use crate::parser::core::SyntaxError;
-use crate::source::Location;
-use crate::source::SourceChar;
 use crate::syntax::*;
 use std::convert::TryFrom;
 
@@ -45,102 +41,6 @@ pub fn is_token_delimiter_char(c: char) -> bool {
 }
 
 impl Lexer {
-    /// Parses a single-quoted string.
-    ///
-    /// The opening `'` must have been consumed before calling this function.
-    /// The closing `'` is consumed in this function.
-    ///
-    /// `opening_location` should be the location of the opening `'`. It is used
-    /// to construct an error value, but this function does not check if it
-    /// actually is a location of `'`.
-    async fn single_quote(&mut self, opening_location: Location) -> Result<WordUnit> {
-        let mut content = String::new();
-        loop {
-            match self.consume_char_if(|_| true).await? {
-                Some(&SourceChar { value: '\'', .. }) => return Ok(SingleQuote(content)),
-                Some(&SourceChar { value, .. }) => content.push(value),
-                None => {
-                    let cause = SyntaxError::UnclosedSingleQuote { opening_location }.into();
-                    let location = self.location().await?.clone();
-                    return Err(Error { cause, location });
-                }
-            }
-        }
-    }
-
-    /// Parses a double-quoted string.
-    ///
-    /// The opening `"` must have been consumed before calling this function.
-    /// The closing `"` is consumed in this function.
-    ///
-    /// `opening_location` should be the location of the opening `"`. It is used
-    /// to construct an error value, but this function does not check if it
-    /// actually is a location of `"`.
-    async fn double_quote(&mut self, opening_location: Location) -> Result<WordUnit> {
-        fn is_delimiter(c: char) -> bool {
-            c == '"'
-        }
-        fn is_escapable(c: char) -> bool {
-            matches!(c, '$' | '`' | '"' | '\\')
-        }
-
-        let content = self.text(is_delimiter, is_escapable).await?;
-
-        if self.skip_if(|c| c == '"').await? {
-            Ok(DoubleQuote(content))
-        } else {
-            let cause = SyntaxError::UnclosedDoubleQuote { opening_location }.into();
-            let location = self.location().await?.clone();
-            Err(Error { cause, location })
-        }
-    }
-
-    /// Parses a word unit.
-    ///
-    /// `is_delimiter` is a function that decides a character is a delimiter. An
-    /// unquoted character is parsed only if `is_delimiter` returns false for it.
-    pub async fn word_unit<F>(&mut self, is_delimiter: F) -> Result<Option<WordUnit>>
-    where
-        F: FnOnce(char) -> bool,
-    {
-        // TODO Parse line continuations before the word unit
-        // TODO Parse other types of word units
-        match self.consume_char_if(|c| c == '\'' || c == '"').await? {
-            None => Ok(self.text_unit(is_delimiter, |_| true).await?.map(Unquoted)),
-            Some(sc) => {
-                let location = sc.location.clone();
-                match sc.value {
-                    '\'' => self.single_quote(location).await.map(Some),
-                    '"' => self.double_quote(location).await.map(Some),
-                    _ => unreachable!(),
-                }
-            }
-        }
-    }
-
-    /// Parses a word token.
-    ///
-    /// `is_delimiter` is a function that decides which character is a delimiter.
-    /// The word ends when an unquoted delimiter is found. To parse a normal word
-    /// token, you should pass [`is_token_delimiter_char`] as `is_delimiter`.
-    /// Other functions can be passed to parse a word that ends with different
-    /// delimiters.
-    ///
-    /// This function does not parse any tilde expansions in the word.
-    /// To parse them, you need to call [`Word::parse_tilde_front`] or
-    /// [`Word::parse_tilde_everywhere`] on the resultant word.
-    pub async fn word<F>(&mut self, mut is_delimiter: F) -> Result<Word>
-    where
-        F: FnMut(char) -> bool,
-    {
-        let location = self.location().await?.clone();
-        let mut units = vec![];
-        while let Some(unit) = self.word_unit(&mut is_delimiter).await? {
-            units.push(unit)
-        }
-        Ok(Word { units, location })
-    }
-
     /// Determines the token ID for the word.
     ///
     /// This is a helper function used by [`Lexer::token`] and does not support
@@ -193,248 +93,13 @@ mod dollar;
 mod heredoc;
 pub mod keyword;
 mod text;
+mod word;
 
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::parser::core::ErrorCause;
     use crate::source::Source;
     use futures::executor::block_on;
-
-    #[test]
-    fn lexer_word_unit_unquoted() {
-        let mut lexer = Lexer::with_source(Source::Unknown, "$()");
-        let result =
-            block_on(lexer.word_unit(|c| panic!("unexpected call to is_delimiter({:?})", c)))
-                .unwrap()
-                .unwrap();
-        if let Unquoted(CommandSubst { content, location }) = result {
-            assert_eq!(content, "");
-            assert_eq!(location.column.get(), 1);
-        } else {
-            panic!("unexpected result {:?}", result);
-        }
-
-        assert_eq!(block_on(lexer.peek_char()), Ok(None));
-    }
-
-    #[test]
-    fn lexer_word_unit_unquoted_escapes() {
-        // Any characters can be escaped in this context.
-        block_on(async {
-            let mut lexer = Lexer::with_source(Source::Unknown, r#"\a\$\`\"\\\'\#"#);
-            let result = lexer
-                .word_unit(|c| panic!("unexpected call to is_delimiter({:?})", c))
-                .await;
-            assert_eq!(result, Ok(Some(Unquoted(Backslashed('a')))));
-            let result = lexer
-                .word_unit(|c| panic!("unexpected call to is_delimiter({:?})", c))
-                .await;
-            assert_eq!(result, Ok(Some(Unquoted(Backslashed('$')))));
-            let result = lexer
-                .word_unit(|c| panic!("unexpected call to is_delimiter({:?})", c))
-                .await;
-            assert_eq!(result, Ok(Some(Unquoted(Backslashed('`')))));
-            let result = lexer
-                .word_unit(|c| panic!("unexpected call to is_delimiter({:?})", c))
-                .await;
-            assert_eq!(result, Ok(Some(Unquoted(Backslashed('"')))));
-            let result = lexer
-                .word_unit(|c| panic!("unexpected call to is_delimiter({:?})", c))
-                .await;
-            assert_eq!(result, Ok(Some(Unquoted(Backslashed('\\')))));
-            let result = lexer
-                .word_unit(|c| panic!("unexpected call to is_delimiter({:?})", c))
-                .await;
-            assert_eq!(result, Ok(Some(Unquoted(Backslashed('\'')))));
-            let result = lexer
-                .word_unit(|c| panic!("unexpected call to is_delimiter({:?})", c))
-                .await;
-            assert_eq!(result, Ok(Some(Unquoted(Backslashed('#')))));
-
-            assert_eq!(lexer.peek_char().await, Ok(None));
-        })
-    }
-
-    #[test]
-    fn lexer_word_unit_single_quote_empty() {
-        let mut lexer = Lexer::with_source(Source::Unknown, "''");
-        let result =
-            block_on(lexer.word_unit(|c| panic!("unexpected call to is_delimiter({:?})", c)))
-                .unwrap()
-                .unwrap();
-        if let SingleQuote(content) = result {
-            assert_eq!(content, "");
-        } else {
-            panic!("unexpected result {:?}", result);
-        }
-
-        assert_eq!(block_on(lexer.peek_char()), Ok(None));
-    }
-
-    #[test]
-    fn lexer_word_unit_single_quote_nonempty() {
-        let mut lexer = Lexer::with_source(Source::Unknown, "'abc\n$def\\'");
-        let result =
-            block_on(lexer.word_unit(|c| panic!("unexpected call to is_delimiter({:?})", c)))
-                .unwrap()
-                .unwrap();
-        if let SingleQuote(content) = result {
-            assert_eq!(content, "abc\n$def\\");
-        } else {
-            panic!("unexpected result {:?}", result);
-        }
-
-        assert_eq!(block_on(lexer.peek_char()), Ok(None));
-    }
-
-    #[test]
-    fn lexer_word_unit_single_quote_unclosed() {
-        let mut lexer = Lexer::with_source(Source::Unknown, "'abc\ndef\\");
-
-        let e = block_on(lexer.word_unit(|c| panic!("unexpected call to is_delimiter({:?})", c)))
-            .unwrap_err();
-        if let ErrorCause::Syntax(SyntaxError::UnclosedSingleQuote { opening_location }) = e.cause {
-            assert_eq!(opening_location.line.value, "'abc\n");
-            assert_eq!(opening_location.line.number.get(), 1);
-            assert_eq!(opening_location.line.source, Source::Unknown);
-            assert_eq!(opening_location.column.get(), 1);
-        } else {
-            panic!("unexpected error cause {:?}", e);
-        }
-        assert_eq!(e.location.line.value, "def\\");
-        assert_eq!(e.location.line.number.get(), 2);
-        assert_eq!(e.location.line.source, Source::Unknown);
-        assert_eq!(e.location.column.get(), 5);
-    }
-
-    #[test]
-    fn lexer_word_unit_double_quote_empty() {
-        let mut lexer = Lexer::with_source(Source::Unknown, "\"\"");
-        let result =
-            block_on(lexer.word_unit(|c| panic!("unexpected call to is_delimiter({:?})", c)))
-                .unwrap()
-                .unwrap();
-        if let DoubleQuote(Text(content)) = result {
-            assert_eq!(content, []);
-        } else {
-            panic!("unexpected result {:?}", result);
-        }
-
-        assert_eq!(block_on(lexer.peek_char()), Ok(None));
-    }
-
-    #[test]
-    fn lexer_word_unit_double_quote_non_empty() {
-        let mut lexer = Lexer::with_source(Source::Unknown, "\"abc\"");
-        let result =
-            block_on(lexer.word_unit(|c| panic!("unexpected call to is_delimiter({:?})", c)))
-                .unwrap()
-                .unwrap();
-        if let DoubleQuote(Text(content)) = result {
-            assert_eq!(content, [Literal('a'), Literal('b'), Literal('c')]);
-        } else {
-            panic!("unexpected result {:?}", result);
-        }
-
-        assert_eq!(block_on(lexer.peek_char()), Ok(None));
-    }
-
-    #[test]
-    fn lexer_word_unit_double_quote_escapes() {
-        // Only the following can be escaped in this context: $ ` " \
-        block_on(async {
-            let mut lexer = Lexer::with_source(Source::Unknown, r#""\a\$\`\"\\\'\#""#);
-            let result = lexer
-                .word_unit(|c| match c {
-                    'a' | '\'' | '#' => true,
-                    _ => panic!("unexpected call to is_delimiter({:?})", c),
-                })
-                .await
-                .unwrap()
-                .unwrap();
-            if let DoubleQuote(Text(ref units)) = result {
-                assert_eq!(
-                    units,
-                    &[
-                        Literal('\\'),
-                        Literal('a'),
-                        Backslashed('$'),
-                        Backslashed('`'),
-                        Backslashed('"'),
-                        Backslashed('\\'),
-                        Literal('\\'),
-                        Literal('\''),
-                        Literal('\\'),
-                        Literal('#'),
-                    ]
-                );
-            } else {
-                panic!("Not a double quote: {:?}", result);
-            }
-
-            assert_eq!(lexer.peek_char().await, Ok(None));
-        })
-    }
-
-    #[test]
-    fn lexer_word_unit_double_quote_unclosed() {
-        let mut lexer = Lexer::with_source(Source::Unknown, "\"abc\ndef");
-
-        let e = block_on(lexer.word_unit(|c| panic!("unexpected call to is_delimiter({:?})", c)))
-            .unwrap_err();
-        if let ErrorCause::Syntax(SyntaxError::UnclosedDoubleQuote { opening_location }) = e.cause {
-            assert_eq!(opening_location.line.value, "\"abc\n");
-            assert_eq!(opening_location.line.number.get(), 1);
-            assert_eq!(opening_location.line.source, Source::Unknown);
-            assert_eq!(opening_location.column.get(), 1);
-        } else {
-            panic!("unexpected error cause {:?}", e);
-        }
-        assert_eq!(e.location.line.value, "def");
-        assert_eq!(e.location.line.number.get(), 2);
-        assert_eq!(e.location.line.source, Source::Unknown);
-        assert_eq!(e.location.column.get(), 4);
-    }
-
-    #[test]
-    fn lexer_word_nonempty() {
-        let mut lexer = Lexer::with_source(Source::Unknown, r"0$(:)X\#");
-        let word = block_on(lexer.word(|_| false)).unwrap();
-        assert_eq!(word.units.len(), 4);
-        assert_eq!(word.units[0], WordUnit::Unquoted(TextUnit::Literal('0')));
-        if let WordUnit::Unquoted(TextUnit::CommandSubst { content, location }) = &word.units[1] {
-            assert_eq!(content, ":");
-            assert_eq!(location.line.value, r"0$(:)X\#");
-            assert_eq!(location.line.number.get(), 1);
-            assert_eq!(location.line.source, Source::Unknown);
-            assert_eq!(location.column.get(), 2);
-        } else {
-            panic!("unexpected word unit: {:?}", word.units[1]);
-        }
-        assert_eq!(word.units[2], WordUnit::Unquoted(TextUnit::Literal('X')));
-        assert_eq!(
-            word.units[3],
-            WordUnit::Unquoted(TextUnit::Backslashed('#'))
-        );
-        assert_eq!(word.location.line.value, r"0$(:)X\#");
-        assert_eq!(word.location.line.number.get(), 1);
-        assert_eq!(word.location.line.source, Source::Unknown);
-        assert_eq!(word.location.column.get(), 1);
-
-        assert_eq!(block_on(lexer.peek_char()), Ok(None));
-    }
-
-    #[test]
-    fn lexer_word_empty() {
-        let mut lexer = Lexer::with_source(Source::Unknown, "");
-        let word = block_on(lexer.word(|_| panic!("unexpected call to is_delimiter"))).unwrap();
-        assert_eq!(word.units, []);
-        assert_eq!(word.location.line.value, "");
-        assert_eq!(word.location.line.number.get(), 1);
-        assert_eq!(word.location.line.source, Source::Unknown);
-        assert_eq!(word.location.column.get(), 1);
-    }
 
     #[test]
     fn lexer_token_empty() {
